Leetcode_medium/src/main/java/com/leetcode/AddTwoNumbers.java
package com.leetcode;

public class AddTwoNumbers {

    /* Time complexity = O(max(n, m))
        Space complexity - O(1) */

    public static ListNode addTwoNumbers(ListNode l1, ListNode l2) {
        ListNode prev = new ListNode(0);
        ListNode current = prev;
        int carry = 0;

        while (l1 != null || l2 != null || carry != 0) {
            int x = l1 == null ? 0 : l1.val;
            int y = l2 == null ? 0 : l2.val;
            int sum = x + y + carry;

            carry = sum / 10;
            current.next = new ListNode(sum % 10);
            current = current.next;

            if (l1 != null) l1 = l1.next;
            if (l2 != null) l2 = l2.next;
        }
        return prev.next;
    }
}
